squashing commit messages

+++ tests/cli/sst_output_dir_json-15.0.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
 #EXT_TEST TEST_FILE_PARAM ALL
-#EXT_TEST TEST_FILE_DESC "Tests basic directory creation using --output-directory"
+#EXT_TEST TEST_FILE_DESC "Tests relocation of JSON config files"
 #EXT_TEST DEP "coreTestElement.coreTestComponent"
+#EXT_TEST TIMEOUT 240
 
 check_odir() {
     odir=$1
     rm -Rf $odir
-    sst --output-directory=$odir cli-sdl.py
+    sst --output-directory=$odir --output-json=sst_output_dir.json cli-sdl.py
     retVal=$?
     if [ $retVal -ne 0 ]; then
         echo "ERROR : $retVal"
@@ -16,11 +17,15 @@ check_odir() {
         echo "ERROR : Output directory ${odir} does not exist!"
         exit 1
     fi
+    if [ ! -e "$odir/sst_output_dir.json" ]; then
+        echo "ERROR missing checkpoint file $odir/sst_output_dir.json"
+        exit 2
+    fi
     rm -Rf $odir
     echo "${odir} OK"
 }
 
-tname=sst_output_dir
+tname=sst_output_dir_json
 
 # Check 1: 1 level subdirectory
 check_odir $tname.$$
@@ -29,7 +34,7 @@ check_odir $tname.$$
 check_odir ./$tname.$$
 
 # Check 3: Absolute path
-check_odir "$(realpath .)/$tname.$$"
+check_odir "$(readlink -f .)/$tname.$$"
 
 # Check 4: 2 level subdirectory
 check_odir "$tname.$$/level2"
